Fixed relinearization error due to precision loss when converting BigInt to Int. Provided study class for better polynomial selection

+++ src/scheme/bfv/KeyGenerator.java
@@ -114,7 +114,7 @@ public class KeyGenerator {
 
         }
 
-        this.relinearizationKeys = new RelinearizationKeys(base.intValue(), keyTuples);
+        this.relinearizationKeys = new RelinearizationKeys(base, keyTuples);
     }
 
     public SecretKey getSecretKey() {
